Fix issue #808: Tooltips don't work on initial hover.
Tooltips don't work on initial hover:

Invoking the tipsy tooltip plugin on a DOM element binds event listeners but
does not trigger the tooltip immediately. This is why the tooltip only
works the second time you hover over something. I fixed this by
triggering tipsy tooltips manually.

Tooltips appear on the left in Firefox:

The tipsy tooltip jQuery plugin had an error calculating the dimensions
of the SVG image element in Firefox which caused all tooltips to appear
on the left side of the window. I added SvgDomWrapper to
provide better compatibility with jQuery plugins that expect HTML
DOM elements.