Issue #809: Allow selected tooltips to be permanently displayed.
Added a new region option called "stickyTooltip" in sites.json
which will keep the marker tooltip permanently visible when true.

This change also upgrades the esri API from 3.3 to 3.8. I needed to
upgrade to use a method that was recently added called getNode() which
gives access to the esri.Graphic DOM element.

I also updated the tipsy jQuery plugin to the latest version and fixed a
bug which prevented the tooltip tails from displaying.